reorganize endpoints
