--- schema/url/categorical-counts.raml
@@ -1,19 +1,0 @@
-#%RAML 1.0 Library
-
-types:
-  
-  CategoricalCountsPostRequest:
-    additionalProperties: false
-    type: VisualizationRequestBase
-    properties:
-      config: CategoricalCountsSpec
-
-  CategoricalCountsSpec:
-    additionalProperties: false
-    properties:
-      variable: VariableSpec
-
-  CategoricalCountsPostResponse:
-    additionalProperties: false
-    properties:
-      categories: BinRangeWithValue[]
